requirements.txt
aws-cdk-lib>=2.241.0,<3.0.0
constructs>=10.5.0,<11.0.0
PyYAML>=6.0
pydantic>=2.0
pydantic-settings>=2.0
